#66 編集した画像が表示できないバグを修正し、軽微なリファクタリングも行った

+++ app/Http/Controllers/Admin/ArticleController.php
@@ -22,11 +22,11 @@ class ArticleController extends Controller
         $form = $request->all();
 
         // フォームに画像があれば画像を保存する処理を行う
-        if (empty($form['image'])) {
+        if (empty($form['image_file'])) {
             $articles->image_path = null;
         } else {
             // 画像ファイルを取得
-            $posted_image = $request->file('image');
+            $posted_image = $request->file('image_file');
 
             // 画像をリサイズしてjpgにencodeする(InterventionImageのImageファサードを使用)
             $resized_image = Image::make($posted_image)->fit(640, 360)->encode('jpg');
@@ -52,7 +52,7 @@ class ArticleController extends Controller
 
         // フォームから送信されてきた_tokenとimageを削除
         unset($form['_token']);
-        unset($form['image']);
+        unset($form['image_file']);
 
         // データベースに保存する
         $articles->fill($form);
@@ -65,7 +65,7 @@ class ArticleController extends Controller
     // editアクションを定義
     public function edit(Request $request)
     {
-        // idで検索してModelからデータを取得する(無ければ404を返す)
+        // idで検索してModelからデータを取得してViewに渡す(無ければ404エラーを返す)
         $articles = Article::find($request->id);
         if (empty($articles)) {
             abort(404);
@@ -84,14 +84,15 @@ class ArticleController extends Controller
         // フォームに画像があれば画像を保存する処理を行う
         // (可読性を考え画像を削除する場合の処理を先にした)
         if (strcmp($request->remove, 'true') == 0) {
-            // 画像を削除
-            Storage::disk('s3')->delete('article_images/' . $articles->image_path);
-            // Storage::delete('public/image/' . $articles->image_path); // ローカル環境での記述
+            // 画像があれば削除する
+            if (isset($articles->image_path)) {
+                Storage::disk('s3')->delete('article_images/' . $articles->image_path);
+            }
             $articles->image_path = null;
-        } elseif (isset($form['image'])) {
+        } elseif (isset($form['image_file'])) {
 
             // 新しい画像ファイルとファイル名を取得
-            $posted_image = $request->file('image');
+            $posted_image = $request->file('image_file');
 
             // 画像をリサイズしてjpgにencodeする(InterventionImageのImageファサードを使用)
             $resized_image = Image::make($posted_image)->fit(640, 360)->encode('jpg');
@@ -103,18 +104,19 @@ class ArticleController extends Controller
             $image_hash = md5($resized_image->__toString());
             $image_name = "{$image_hash}.jpg";
 
-            // 現在設定中の画像を削除し、加工した新しい画像を保存する
-            Storage::disk('s3')->delete('article_images/' . $articles->image_path);
-            Storage::disk('s3')->put('article_images/' . $image_name, $resized_image);
-            // Storage::put('public/image/' . $image_name, $resized_image); // ローカル環境での記述
+            // 現在設定中の画像があれば削除し、加工した新しい画像を保存する
+            if (isset($articles->image_path)) {
+                Storage::disk('s3')->delete('article_images/' . $articles->image_path);
+            }
+            Storage::disk('s3')->put('article_images/' . $image_name, $resized_image, 'public');
 
-            // S3のファイルのURLを取得してuser_image_pathカラムに書き込む
+            // S3の新しい画像ファイルのURLを取得してuser_image_pathカラムに書き込む
             $articles->image_path = Storage::disk('s3')->url('article_images/' . $image_name);
         }
 
         // フォームの不要なデータを削除する
         unset($form['_token']);
-        unset($form['image']);
+        unset($form['image_file']);
         unset($form['remove']);
 
         // 編集日時を追加する
@@ -131,9 +133,12 @@ class ArticleController extends Controller
     {
         // Modelからデータを取得して削除(投稿idで検索)
         $articles = Article::find($request->id);
-        Storage::disk('s3')->delete('user_images/' . $articles->image_path); // S3のファイルを削除
-        // Storage::delete('public/image/' . $articles->image_path); // ローカル環境での記述
-        $articles->delete(); // テーブルのファイルパスを削除
+
+        // 画像パスがあればS3の画像ファイルを削除し、テーブルのデータを削除する
+        if (isset($articles->image_path)) {
+            Storage::disk('s3')->delete('user_images/' . $articles->image_path);
+        }
+        $articles->delete();
         return redirect('admin/articles');
     }
